Update git-core to v2.55.0

+++ src/store.rs
@@ -10,7 +10,7 @@ use std::hash::Hash;
 use std::io::{copy, BufRead, BufReader, Read, Write};
 use std::iter::{repeat, IntoIterator};
 use std::num::NonZeroU32;
-use std::os::raw::{c_char, c_int, c_ulong};
+use std::os::raw::{c_char, c_int};
 use std::process::{Command, Stdio};
 use std::sync::Mutex;
 use std::{mem, ptr};
@@ -1440,7 +1440,7 @@ extern "C" {
     );
     pub fn do_set_replace(replaced: *const object_id, replace_with: *const object_id);
     fn get_object_entry(oid: *const object_id) -> *const object_entry;
-    fn unpack_object_entry(oe: *const object_entry, buf: *mut *mut c_char, len: *mut c_ulong);
+    fn unpack_object_entry(oe: *const object_entry, buf: *mut *mut c_char, len: *mut usize);
 }
 
 pub fn store_git_blob(blob_buf: &[u8]) -> BlobId {
@@ -1467,7 +1467,7 @@ pub fn store_git_tree(tree_buf: &[u8], reference: Option<TreeId>) -> TreeId {
                 let mut reftree_buf = ptr::null_mut();
                 let mut len = 0;
                 unpack_object_entry(oe, &mut reftree_buf, &mut len);
-                ref_tree = Some(FfiBox::from_raw_parts(reftree_buf as *mut _, len as usize));
+                ref_tree = Some(FfiBox::from_raw_parts(reftree_buf as *mut _, len));
             }
         }
         let mut result = object_id::default();
